[change] 評価値がインフレしすぎないように最初から1/10にした。

+++ book_evaluation.h
@@ -3,6 +3,6 @@
 
 #include "const.h"
 
-extern int evaluation(char board[BOARD_SIZE+2][BOARD_SIZE+2], int player);
+extern double evaluation(char board[BOARD_SIZE+2][BOARD_SIZE+2], int player);
 
 #endif
